location_integration_layer/requirements.txt
# Main Dagster version 
dagster==1.6.3
dagster-webserver==1.6.3
dagster-graphql==1.6.3

# Dagster integrations
dagster-docker==0.22.3
dagster-postgres==0.22.3
dagster-duckdb==0.22.3
dagster-duckdb-pandas==0.22.3
dagster-dbt==0.22.3
dagster-gcp==0.22.3

# Duck DB
duckdb==0.9.2

# dbt 
dbt-duckdb==1.7.0

# Other 
pandas
pydantic
google-cloud-storage